refactor(decorators): move Container to new component
